// cloudflare-worker/src/data/spatialFacts.generated.ts
export const SPATIAL_FACTS = {
  "parcelId": "KR9B_117A55",
  "caseId": "KR9B_117A55",
  "resolvedFrom": {
    "kind": "address",
    "value": "KR 9B #117A-55",
    "confidence": "high"
  },
  "generatedAt": "2026-07-06T01:32:05.823407+00:00",
  "spatialReference": "EPSG:4686",
  "parcel": {
    "address": "KR 9B #117A-55",
    "lotCode": "008415053027",
    "frontPoint": {
      "x": -74.03494603000001,
      "y": 4.696439515154202
    },
    "centroid": {
      "x": -74.03506977833872,
      "y": 4.696509265259457
    }
  },
// ...
} satisfies GeneratedSpatialFacts